pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Page Analyzer is a site that analyzes specified pages for SEO suitability."
authors = ["Tata Pestova <tata.pestova@inbox.ru>"]
readme = "README.md"
repository = "https://github.com/tatapestova/python-project-83"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Flask",
    "Programming Language :: Python :: 3.10",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application"
]
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
gunicorn = "^20.1.0"
flask = "^2.3.2"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
validators = "^0.20.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
